backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
psycopg[binary]==3.2.3
supabase==2.6.0
openai==1.50.2
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
numpy==2.1.2

